Refactor SampleLoader class back into procedural functions
